[medium] fixes to the agent upgrade protocol to support initialization by service

+++ src/mig/modules/agentdestroy/agentdestroy.go
@@ -43,7 +43,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"regexp"
+	"os/exec"
 	"runtime"
 )
 
@@ -73,9 +73,8 @@ func (p Parameters) Validate() (err error) {
 	if p.PID < 2 || p.PID > 65535 {
 		return fmt.Errorf("PID '%s' is not in the range [2:65535]", p.PID)
 	}
-	versionre := regexp.MustCompile(`^[a-z0-9]{7}-[0-9]{12}$`)
-	if !versionre.MatchString(p.Version) {
-		return fmt.Errorf("Parameter 'bin_version' with value '%s' is invalid. Expecting version.", p.Version)
+	if p.Version == "" {
+		return fmt.Errorf("parameter 'version' is empty. Expecting version.")
 	}
 	return
 }
@@ -95,45 +94,74 @@ func Run(Args []byte) string {
 	}
 	// Refuse to suicide
 	if params.PID == os.Getppid() {
-		results.Success = false
 		results.Error = fmt.Sprintf("PID '%d' is mine. Refusing to suicide.", params.PID)
 		return buildResults(results)
 	}
 
-	// First kill the agent PID
-	results.Success = true
+	// get the path of the agent's executable
+	var binary string
+	switch runtime.GOOS {
+	case "linux", "darwin", "freebsd", "openbsd", "netbsd":
+		binary, err = os.Readlink(fmt.Sprintf("/proc/%d/exe", params.PID))
+		if err != nil {
+			results.Error = fmt.Sprintf("Binary path of PID '%d' not found: '%v'", params.PID, err)
+			return buildResults(results)
+		}
+	case "windows":
+		binary = fmt.Sprintf("C:/Windows/mig-agent-%s.exe", params.Version)
+	default:
+		results.Error = fmt.Sprintf("'%s' isn't a supported OS", runtime.GOOS)
+		return buildResults(results)
+	}
+
+	// check that the binary we're removing has the right version
+	version, err := getAgentVersion(binary)
+	if err != nil {
+		results.Error = fmt.Sprintf("Failed to check agent version: '%v'", err)
+		return buildResults(results)
+	}
+	if version != params.Version {
+		results.Error = fmt.Sprintf("Version mismatch. Expected '%s', found '%s'", params.Version, version)
+		return buildResults(results)
+	}
+	err = os.Remove(binary)
+	if err != nil {
+		results.Error = fmt.Sprintf("Failed to remove binary '%s': '%v'", binary, err)
+		return buildResults(results)
+	}
+
+	// Then kill the PID
 	process, err := os.FindProcess(params.PID)
 	if err != nil {
 		results.Error = fmt.Sprintf("PID '%d' not found. Returned error '%v'", params.PID, err)
-		results.Success = false
 		return buildResults(results)
 	} else {
 		err = process.Kill()
 		if err != nil {
 			results.Error = fmt.Sprintf("PID '%d' not killed. Returned error '%v'", params.PID, err)
-			results.Success = false
 			return buildResults(results)
 		}
 	}
+	results.Success = true
+	return buildResults(results)
+}
 
-	// Then remove the agent binary
-	var binary string
-	switch runtime.GOOS {
-	case "linux", "darwin", "freebsd", "openbsd", "netbsd":
-		binary = fmt.Sprintf("/sbin/mig-agent-%s", params.Version)
-	case "windows":
-		binary = fmt.Sprintf("C:/Windows/mig-agent-%s.exe", params.Version)
-	default:
-		results.Error = fmt.Sprintf("'%s' isn't a supported OS", runtime.GOOS)
-		results.Success = false
-		return buildResults(results)
-	}
-	err = os.Remove(binary)
+// Run the agent binary to obtain its version
+func getAgentVersion(binary string) (cversion string, err error) {
+	defer func() {
+		if e := recover(); e != nil {
+			err = fmt.Errorf("getAgentVersion() -> %v", e)
+		}
+	}()
+	out, err := exec.Command(binary, "-V").Output()
 	if err != nil {
 		panic(err)
 	}
-
-	return buildResults(results)
+	if len(out) < 2 {
+		panic("Failed to retrieve agent version.")
+	}
+	cversion = string(out[:len(out)-1])
+	return
 }
 
 func buildResults(results Results) (jsonResults string) {
